+++ Tests/HummingbirdTests/ApplicationTests.swift
@@ -557,7 +557,7 @@ final class ApplicationTests: XCTestCase {
     /// test we can create an application that accepts a responder with an empty context
     func testEmptyRequestContext() async throws {
         struct EmptyRequestContext: InitializedFrom {
-            typealias Source = ServerRequestContextSource
+            typealias Source = ApplicationRequestContextSource
             init(source: Source) {}
         }
         let app = Application(
